content_creation_agency/requirements.txt
agency-swarm
openai
python-dotenv
tavily-python
nltk
pytrends
requests 